requirements.txt
fastapi
uvicorn[standard]
asyncpg
sqlalchemy>=2.0
alembic

pydantic>=2.0
pydantic-settings>=2.0

python-jose[cryptography]
httpx
aiohttp
aiofiles
jinja2
python-multipart

pytest
pytest-asyncio
pytest-cov